+++ tests/unit/utils/open_url_framework.py
@@ -58,50 +58,17 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 
-import json
-import traceback
-
-import pytest
-
 from mock import MagicMock
 
-from ansible.module_utils._text import to_native
 from ansible.module_utils.six.moves.urllib.error import HTTPError
-from ansible.module_utils.six.moves.urllib.parse import parse_qs
+
+from ._utils import (
+    CallBase as _CallBase,
+    validate_call as _validate_call,
+)
 
 
-def _extract_query(url):
-    query_index = url.find('?')
-    fragment_index = url.find('#')
-    if query_index > fragment_index and fragment_index > 0:
-        query_index = -1
-    if query_index < 0:
-        return {}
-    if fragment_index > 0:
-        query = url[query_index + 1:fragment_index]
-    else:
-        query = url[query_index + 1:]
-    return parse_qs(query, keep_blank_values=True)
-
-
-def _reduce_url(url, remove_query=False, remove_fragment=False):
-    fragment_index = url.find('#')
-    if remove_fragment and fragment_index > 0:
-        url = url[:fragment_index]
-        fragment_index = -1
-    if remove_query:
-        query_index = url.find('?')
-        if query_index > fragment_index and fragment_index > 0:
-            query_index = -1
-        if query_index > 0:
-            if fragment_index > 0:
-                url = url[:query_index] + url[fragment_index:]
-            else:
-                url = url[:query_index]
-    return url
-
-
-class OpenUrlCall:
+class OpenUrlCall(_CallBase):
     '''
     Describes one call to ``open_url()``.
 
@@ -114,29 +81,8 @@ class OpenUrlCall:
         Create a ``open_url()`` expected call. Must be passed information on the expected
         HTTP method and the HTTP status returned by the call.
         '''
-        assert method == method.upper(), \
-            'HTTP method names are case-sensitive and should be upper-case (RFCs 7230 and 7231)'
-        self.method = method
-        self.status = status
+        super(OpenUrlCall, self).__init__(method, status)
         self.exception_generator = None
-        self.body = None
-        self.headers = {}
-        self.error_data = {}
-        self.expected_url = None
-        self.expected_url_without_query = False
-        self.expected_url_without_fragment = False
-        self.expected_headers = {}
-        self.expected_query = {}
-        self.expected_content = None
-        self.expected_content_predicate = None
-        self.form_parse = False
-        self.form_present = set()
-        self.form_values = {}
-        self.form_values_one = {}
-        self.json_parse = False
-        self.json_present = set()
-        self.json_absent = set()
-        self.json_values = {}
 
     def exception(self, exception_generator):
         '''
@@ -152,144 +98,18 @@ class OpenUrlCall:
         '''
         Builder method to set return body of the ``open_url()`` call. Must be a bytes string.
         '''
-        self.body = body
-        assert self.error_data.get('body') is None, 'Error body must not be given if body is provided'
         assert self.exception_generator is None, 'Exception generator must not be given if body is provided'
-        return self
-
-    def result_str(self, str_body):
-        '''
-        Builder method to set return body of the ``open_url()`` call as a text string.
-        '''
-        return self.result(str_body.encode('utf-8'))
-
-    def result_json(self, json_body):
-        '''
-        Builder method to set return body of the ``open_url()`` call as a JSON object.
-        '''
-        return self.result(json.dumps(json_body).encode('utf-8'))
+        return super(OpenUrlCall, self).result(body)
 
     def result_error(self, body=None):
         '''
         Builder method to set return body of the ``open_url()`` call in case of an error.
         '''
-        self.error_data['body'] = body
-        if body is not None:
-            assert self.body is None, 'Result must not be given if error body is provided'
         assert self.exception_generator is None, 'Exception generator must not be given if error is provided'
-        return self
-
-    def expect_url(self, url, without_query=False, without_fragment=False):
-        '''
-        Builder method to set the expected URL for the ``open_url()`` call.
-        '''
-        self.expected_url = url
-        self.expected_url_without_query = without_query
-        self.expected_url_without_fragment = without_fragment
-        return self
-
-    def expect_query_values(self, parameter, *values):
-        '''
-        Builder method to set an expected query parameter for the ``open_url()`` call.
-        '''
-        self.expected_query[parameter] = list(values)
-        return self
-
-    def return_header(self, name, value):
-        '''
-        Builder method to set a returned header for the ``open_url()`` call.
-        '''
-        assert value is not None
-        self.headers[name] = value
-        return self
-
-    def expect_header(self, name, value):
-        '''
-        Builder method to set an expected header value for a ``open_url()`` call.
-        '''
-        self.expected_headers[name] = value
-        return self
-
-    def expect_header_unset(self, name):
-        '''
-        Builder method to set an expected non-set header for a ``open_url()`` call.
-        '''
-        self.expected_headers[name] = None
-        return self
-
-    def expect_content(self, content):
-        '''
-        Builder method to set an expected content for a ``open_url()`` call.
-        '''
-        self.expected_content = content
-        return self
-
-    def expect_content_predicate(self, content_predicate):
-        '''
-        Builder method to set an expected content predicate for a ``open_url()`` call.
-        '''
-        self.expected_content_predicate = content_predicate
-        return self
-
-    def expect_form_present(self, key):
-        '''
-        Builder method to set an expected form field for a ``open_url()`` call.
-        '''
-        assert not self.json_parse
-        self.form_parse = True
-        self.form_present.add(key)
-        return self
-
-    def expect_form_value(self, key, value):
-        '''
-        Builder method to set an expected form value for a ``open_url()`` call.
-        '''
-        assert not self.json_parse
-        self.form_parse = True
-        self.form_values[key] = [value]
-        return self
-
-    def expect_form_value_absent(self, key):
-        '''
-        Builder method to set an expectedly absent form field for a ``open_url()`` call.
-        '''
-        assert not self.json_parse
-        self.form_parse = True
-        self.form_values[key] = []
-        return self
-
-    def expect_json_present(self, key):
-        '''
-        Builder method to set an expected JSON field for a ``open_url()`` call.
-        The key must be a sequence: strings denote fields in objects, integers denote array indices.
-        '''
-        assert not self.form_parse
-        self.json_parse = True
-        self.json_present.add(tuple(key))
-        return self
-
-    def expect_json_value(self, key, value):
-        '''
-        Builder method to set an expected JSON value for a ``open_url()`` call.
-        The key must be a sequence: strings denote fields in objects, integers denote array indices.
-        '''
-        assert not self.form_parse
-        self.json_parse = True
-        self.json_values[tuple(key)] = value
-        return self
-
-    def expect_json_value_absent(self, key):
-        '''
-        Builder method to set an expectedly absent JSON field for a ``open_url()`` call.
-        The key must be a sequence: strings denote fields in objects, integers denote array indices.
-        '''
-        assert not self.form_parse
-        self.json_parse = True
-        self.json_absent.add(tuple(key))
-        return self
+        return super(OpenUrlCall, self).result_error(body=body)
 
 
-class OpenUrlProxy:
+class OpenUrlProxy(object):
     '''
     Proxy for ``open_url()``.
 
@@ -303,110 +123,6 @@ class OpenUrlProxy:
         '''
         self.calls = calls
         self.index = 0
-
-    def _validate_form(self, call, data):
-        '''
-        Validate form contents.
-        '''
-        form = {}
-        if data is not None:
-            form = parse_qs(to_native(data), keep_blank_values=True)
-        for k in call.form_present:
-            assert k in form, 'Form key "{0}" not present'.format(k)
-        for k, v in call.form_values.items():
-            if len(v) == 0:
-                assert k not in form, 'Form key "{0}" not absent'.format(k)
-            else:
-                assert form[k] == v, 'Form key "{0}" has not values {1}, but {2}'.format(k, v, form[k])
-
-    def _validate_form(self, call, data):
-        '''
-        Validate form contents.
-        '''
-        form = {}
-        if data is not None:
-            form = parse_qs(to_native(data), keep_blank_values=True)
-        for k in call.form_present:
-            assert k in form, 'Form key "{0}" not present'.format(k)
-        for k, v in call.form_values.items():
-            if len(v) == 0:
-                assert k not in form, 'Form key "{0}" not absent'.format(k)
-            else:
-                assert form[k] == v, 'Form key "{0}" has not values {1}, but {2}'.format(k, v, form[k])
-
-    def _descend_json(self, data, key):
-        if not key:
-            return data, True
-        for index, k in enumerate(key[:-1]):
-            if isinstance(k, int):
-                if not isinstance(data, (list, tuple)):
-                    raise Exception('Cannot resolve JSON key {0} in data: not a list on last level'.format(self._format_json_key(key[:index + 1])))
-                if key[-1] < 0 or key[-1] >= len(data):
-                    raise Exception('Cannot find JSON key {0} in data: index out of bounds'.format(self._format_json_key(key[:index + 1])))
-            else:
-                if not isinstance(data, dict):
-                    raise Exception('Cannot resolve JSON key {0} in data: not a dictionary on last level'.format(self._format_json_key(key[:index + 1])))
-                if key[-1] not in data:
-                    raise Exception('Cannot find JSON key {0} in data: key not present'.format(self._format_json_key(key[:index + 1])))
-            data = data[k]
-        if isinstance(key[-1], int):
-            if not isinstance(data, (list, tuple)):
-                raise Exception('Cannot resolve JSON key {0} in data: not a list on last level'.format(self._format_json_key(key)))
-            if key[-1] < 0 or key[-1] >= len(data):
-                return None, False
-        else:
-            if not isinstance(data, dict):
-                raise Exception('Cannot resolve JSON key {0} in data: not a dictionary on last level'.format(self._format_json_key(key)))
-            if key[-1] not in data:
-                return None, False
-        return data[key[-1]], True
-
-    def _validate_json(self, call, data):
-        '''
-        Validate JSON contents.
-        '''
-        data = json.loads(to_native(data))
-        for k in call.json_present:
-            dummy, present = self._descend_json(data, k)
-            assert present, 'JSON key {0} not present'.format(self._format_json_key(k))
-        for k in call.json_absent:
-            dummy, present = self._descend_json(data, k)
-            assert not present, 'JSON key {0} present'.format(self._format_json_key(k))
-        for k, v in call.json_values.items():
-            value, present = self._descend_json(data, k)
-            assert present, 'JSON key "{0}" not present'.format(self._format_json_key(k))
-            assert value == v, 'JSON key "{0}" has not value {1!r}, but {2!r}'.format(self._format_json_key(k), v, value)
-
-    def _validate_query(self, call, url):
-        '''
-        Validate query parameters of a call.
-        '''
-        query = _extract_query(url)
-        for k, v in call.expected_query.items():
-            if k not in query:
-                assert query.get(k) == v, \
-                    'Query parameter "{0}" not specified for open_url call'.format(k)
-            else:
-                assert query.get(k) == v, \
-                    'Query parameter "{0}" specified for open_url call, but with wrong value ({1!r} instead of {2!r})'.format(
-                        k, query.get(k), v)
-
-    def _validate_headers(self, call, headers):
-        '''
-        Validate headers of a call.
-        '''
-        given_headers = {}
-        if headers is not None:
-            for k, v in headers.items():
-                given_headers[k.lower()] = v
-        for k, v in call.expected_headers.items():
-            if v is None:
-                assert k.lower() not in given_headers, \
-                    'Header "{0}" specified for open_url call, but should not be'.format(k)
-            else:
-                assert given_headers.get(k.lower()) == v, \
-                    'Header "{0}" specified for fetch_url call, but with wrong value ({1!r} instead of {2!r})'.format(
-                        k, given_headers.get(k.lower()), v)
 
     def __call__(self, url, data=None, headers=None, method=None, use_proxy=True,
                  force=False, last_mod_time=None, timeout=10, validate_certs=True,
@@ -423,33 +139,7 @@ class OpenUrlProxy:
         self.index += 1
 
         # Validate call
-        assert method == call.method, 'Expected method does not match for fetch_url call: {0!r} instead of {1!r}'.format(
-            method, call.method)
-        if call.expected_url is not None:
-            reduced_url = _reduce_url(
-                url,
-                remove_query=call.expected_url_without_query,
-                remove_fragment=call.expected_url_without_fragment)
-            assert reduced_url == call.expected_url, \
-                'Exepected URL does not match for open_url call: {0!r} instead of {1!r}'.format(reduced_url, call.expected_url)
-        if call.expected_query:
-            self._validate_query(call, url)
-        if call.expected_headers:
-            self._validate_headers(call, headers)
-        if call.expected_content is not None:
-            assert data == call.expected_content, 'Expected content does not match for open_url call: {0!r} instead of {1!r}'.format(
-                data, call.expected_content)
-        if call.expected_content_predicate:
-            try:
-                assert call.expected_content_predicate(data), 'Predicate has falsy result'
-            except Exception as e:
-                raise AssertionError(
-                    'Content does not match predicate for open_url call: {0}\n\n{1}'.format(
-                        e, traceback.format_exc()))
-        if call.form_parse:
-            self._validate_form(call, data)
-        if call.json_parse:
-            self._validate_json(call, data)
+        _validate_call(call, method=method, url=url, headers=headers, data=data)
 
         # Compose result
         if call.exception_generator:
